sentinel :: move most processing to protocols to be able to support other types of processing
