+++ src/main/java/emissary/core/sentinel/protocols/actions/Exit.java
@@ -1,6 +1,6 @@
 package emissary.core.sentinel.protocols.actions;
 
-import emissary.core.sentinel.Sentinel;
+import emissary.core.sentinel.protocols.trackers.Tracker;
 
 import java.util.Map;
 
@@ -11,8 +11,8 @@ public class Exit extends Action {
 
     @Override
     @SuppressWarnings("SystemExitOutsideMain")
-    public void trigger(Map<String, Sentinel.Tracker> trackers) {
-        logger.error("Sentinel detected unrecoverable agents, exiting now -- {}", format(trackers));
+    public void trigger(Map<String, Tracker> trackers) {
+        logger.error("Sentinel detected unrecoverable state, exiting now -- {}", format(trackers));
         System.exit(1);
     }
 }
