+++ src/main/java/emissary/core/sentinel/protocols/trackers/Tracker.java
@@ -1,0 +1,5 @@
+package emissary.core.sentinel.protocols.trackers;
+
+public interface Tracker {
+
+}
