+++ src/main/resources/emissary/core/sentinel/protocols/ProtocolSample.cfg
@@ -1,3 +1,4 @@
+PROTOCOL_CLASS = "emissary.core.sentinel.protocols.AgentProtocol"
 ENABLED = false
 
 # Default action is Notify
